Don't attempt to construct file object with null file name, don't try to delete

+++ src/itkach/aard2/DescriptorStore.java
@@ -53,6 +53,9 @@ class DescriptorStore<T extends BaseDescriptor> {
     }
 
     boolean delete(String itemId) {
+        if (itemId == null) {
+            return false;
+        }
         return new File(dir, itemId).delete();
     }
